+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version latest (0.1.7?, 2019-09-??)
 
 - ...
+- enh: replace "lambda" with "underused/burstable {hourly,daily}"
 
 
 Version 0.3.1 (2019-09-16)